requirements.txt
fastapi==0.115.2
starlette==0.38.5
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-telegram-bot==20.8
imageio-ffmpeg==0.5.1
